requirements.txt
Pillow==9.4.0
PyMuPDF==1.23.4
schedule==1.2.1